Moved JS files related to SDEV305 to that folder - Updated file paths to account for change
