+++ 305/js/responsive-accordion-toggle.js
@@ -1,4 +1,4 @@
-import * as helpers from './helpers.js';
+import * as helpers from '../../js/helpers.js';
 
 "use strict";
